API reference

Main API

DarkCurves.curve_point_type — Function

curve_point_type(::Type{<:EllipticCurve})

Returns the type of a curve point (a subtype of EllipticCurvePoint) for an object of type EllipticCurve.

Objects of this type support zero(), iszero(), one(), rand(), +, -, ==, * (by a scalar type).

Base point is returned by one(), and infinity point by zero().

DarkCurves.curve_scalar_type — Function

curve_scalar_type(::Type{<:EllipticCurve})

Returns the type of an associated scalar for an object of type EllipticCurve. All operations with objects of this type are done modulo curve order.

Objects of this type support zero(), iszero(), one(), rand(), +, -, ==, * (by a scalar type or a point type), iseven(), isodd(), >>, inv(), divrem(), trailing_zeros(), DarkIntegers.num_bits().

Abstract types

DarkCurves.EllipticCurve — Type

Abstract type for an elliptic curve "bundle" that includes a point type, a scalar type, and operations on them.

DarkCurves.EllipticCurvePoint — Type

Abstract type of a curve point.

Batch operations

DarkCurves.lin_comb — Function

Calculates a linear combination of curve points given an array of points and an array of coefficients.

DarkCurves.batch_mul — Function

Returns points .* coeff.

Curve implementations

SECP256k1

DarkCurves.Curve_secp256k1 — Type

SECP256k1 curve y^2 = x^3 + 7 over the field of integers modulo 2^256 - 2^32 - 2^9 - 2^8 - 2^7 - 2^6 - 2^4 - 1 with the order (little-endian) 0xFFFFFFFF FFFFFFFF FFFFFFFF FFFFFFFE BAAEDCE6 AF48A03B BFD25E8C D0364141.